--- pyrit/adv_suffix/llm_attacks/gcg/__init__.py
@@ -1,6 +1,0 @@
-from .gcg_attack import GCGAttackPrompt as AttackPrompt
-from .gcg_attack import GCGPromptManager as PromptManager
-from .gcg_attack import GCGMultiPromptAttack as MultiPromptAttack
-
-from pyrit.adv_suffix.llm_attacks.base.attack_manager import ProgressiveMultiPromptAttack
-from pyrit.adv_suffix.llm_attacks.base.attack_manager import IndividualPromptAttack
