+++ pyrit/optimizers/GreedyCoordinateGradient/gcg/__init__.py
@@ -1,0 +1,6 @@
+from .gcg_attack import GCGAttackPrompt as AttackPrompt
+from .gcg_attack import GCGPromptManager as PromptManager
+from .gcg_attack import GCGMultiPromptAttack as MultiPromptAttack
+
+from pyrit.optimizers.GreedyCoordinateGradient.base.attack_manager import ProgressiveMultiPromptAttack
+from pyrit.optimizers.GreedyCoordinateGradient.base.attack_manager import IndividualPromptAttack
