+++ pyrit/optimizers/experiments/configs/individual_mistral.yaml
